+++ Cargo.toml
@@ -1,32 +1,5 @@
-[package]
-name = "gestalt"
-version = "0.0.1"
-authors = ["Emily 'Gyro' C. <gyrocoder@gmail.com>", "trashbyte"]
-edition = "2018"
+[workspace]
 
-[dependencies]
-byteorder = "1.3.4"
-cgmath = "0.18.0"
-crossbeam-channel = "0.5.0"
-custom_error = "1.8.0"
-hashbrown = { version = "0.11.2", features = ["serde"] } 
-lazy_static = "1.4.0"
-log = "0.4.11"
-num = "0.4.0"
-parking_lot = "0.11.0"
-rand = "0.8.4"
-semver = "1.0.4"
-serde = { version = "1.0.108", features = ["derive"] }
-simplelog = "0.10.2"
-uuid = { version = "0.8.1", features = ["serde", "v4"] }
-#winit = "0.25"
-
-# Scripting
-wasmtime = "0.30.0"
-
-# Image loading
-# image = "0.23.8"
-
-# Possible options for cryptography... 
-# snow = "0.8"
-# sodiumoxide = "0.2.6"
+members = [
+    "core",
+]
