+++ core/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "gestalt"
+version = "0.0.1"
+authors = ["Emily 'Gyro' C. <gyrocoder@gmail.com>", "trashbyte"]
+edition = "2018"
+
+[dependencies]
+byteorder = "1.3.4"
+crossbeam-channel = "0.5.0"
+custom_error = "1.8.0"
+hashbrown = { version = "0.11.2", features = ["serde"] } 
+hecs = "0.7.3"
+lazy_static = "1.4.0"
+log = "0.4.11"
+num = "0.4.0"
+parking_lot = "0.11.0"
+rand = "0.8.4"
+semver = "1.0.4"
+serde = { version = "1.0.108", features = ["derive"] }
+sha2 = "0.10.1"
+simplelog = "0.11"
+uuid = { version = "0.8.1", features = ["serde", "v4"] }
+#winit = "0.25"
+
+# Scripting
+mlua = { version = "0.7.1", features = ["lua52"] }
+
+# Image loading
+# image = "0.23.8"
+
+# Possible options for cryptography... 
+# snow = "0.8"
+# sodiumoxide = "0.2.6"
