+++ core/src/common/identity.rs
@@ -1,0 +1,6 @@
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+
+#[repr(C)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize, Hash)]
+pub struct NodeId(Uuid);
