+++ core/src/common/mod.rs
@@ -1,0 +1,5 @@
+#[macro_use]
+pub mod voxelmath;
+
+pub mod identity;
+pub mod resource;
