--- src/common/mod.rs
@@ -1,2 +1,0 @@
-pub mod ecs;
-#[macro_use] pub mod voxelmath;
